reassign edited message instance to 'this'
give devs the 'latest' (reassigned) instance. I think prior to this the reaction cache for rolling could also maybe have been not totally complete (idk how reaction-user fetching influences it exactly)  = now better. ||||| For the record: this is based on test i conducted, since androz had told me the instance would auto update/reassign itself, but that didn't seemed to be the case in my tests. (if smth comes up which changes that conclusion/if theres a problem with the reassigning, lmk)