Usage of functions instead of package
Using functions instead of packages can lead to improved performance. Functions are generally lightweight compared to packages, as they don't involve the overhead of loading external dependencies. If you have a small, self-contained functionality that doesn't require the full capabilities of a package, implementing it as a function can result in faster execution times